refactor functionPointer
