+++ tools/apiview/parsers/rust-api-parser/src/process-items/utils/typeToReviewTokens.ts
@@ -1,5 +1,6 @@
 import { Type } from "../../../rustdoc-types/output/rustdoc-types";
 import { ReviewToken, TokenKind } from "../../models/apiview-models";
+import { processFunctionPointer } from "./processFunctionPointer";
 import { processGenericArgs } from "./processGenerics";
 import { shouldElideLifetime } from "./shouldElideLifeTime";
 
@@ -65,103 +66,7 @@ export function typeToReviewTokens(type: Type): ReviewToken[] {
   } else if ("primitive" in type) {
     return [{ Kind: TokenKind.TypeName, Value: type.primitive, HasSuffixSpace: false }];
   } else if ("function_pointer" in type) {
-    const tokens: ReviewToken[] = [];
-
-    // Add generic parameters (for<'a> etc)
-    if (type.function_pointer.generic_params.length > 0) {
-      tokens.push({ Kind: TokenKind.Keyword, Value: "for", HasSuffixSpace: false });
-      tokens.push({ Kind: TokenKind.Punctuation, Value: "<", HasSuffixSpace: false });
-
-      type.function_pointer.generic_params.forEach((param, index) => {
-        tokens.push({ Kind: TokenKind.Text, Value: param.name, HasSuffixSpace: false });
-
-        if (index < type.function_pointer.generic_params.length - 1) {
-          tokens.push({ Kind: TokenKind.Punctuation, Value: ", ", HasSuffixSpace: false });
-        }
-      });
-
-      tokens.push({ Kind: TokenKind.Punctuation, Value: "> ", HasSuffixSpace: false });
-    }
-
-    // Add function modifiers
-    const header = type.function_pointer.header;
-
-    // Handle unsafe
-    if (header.is_unsafe) {
-      tokens.push({ Kind: TokenKind.Keyword, Value: "unsafe", HasSuffixSpace: true });
-    }
-
-    // Handle extern ABI
-    if (header.abi !== "Rust") {
-      tokens.push({ Kind: TokenKind.Keyword, Value: "extern", HasSuffixSpace: true });
-
-      let abiString = "";
-      if (typeof header.abi === "string") {
-        // For simple ABI strings
-        abiString = header.abi;
-      } else {
-        // For complex ABI objects
-        const abiKey = Object.keys(header.abi)[0];
-        if (abiKey) {
-          abiString = abiKey;
-        }
-      }
-
-      if (abiString && abiString !== "Rust") {
-        tokens.push({ Kind: TokenKind.Text, Value: `"${abiString}"`, HasSuffixSpace: true });
-      }
-    }
-
-    // Handle const
-    if (header.is_const) {
-      tokens.push({ Kind: TokenKind.Keyword, Value: "const", HasSuffixSpace: true });
-    }
-
-    // Handle async
-    if (header.is_async) {
-      tokens.push({ Kind: TokenKind.Keyword, Value: "async", HasSuffixSpace: true });
-    }
-
-    // Add fn keyword
-    tokens.push({ Kind: TokenKind.Keyword, Value: "fn", HasSuffixSpace: false });
-
-    // Add parameters
-    tokens.push({ Kind: TokenKind.Punctuation, Value: "(", HasSuffixSpace: false });
-
-    const signature = type.function_pointer.sig;
-    signature.inputs.forEach(([paramName, paramType], index) => {
-      // Add parameter name if present
-      if (paramName) {
-        tokens.push({ Kind: TokenKind.Text, Value: paramName, HasSuffixSpace: false });
-        tokens.push({ Kind: TokenKind.Punctuation, Value: ": ", HasSuffixSpace: false });
-      }
-
-      // Add parameter type
-      tokens.push(...typeToReviewTokens(paramType));
-
-      // Add comma if not the last parameter
-      if (index < signature.inputs.length - 1) {
-        tokens.push({ Kind: TokenKind.Punctuation, Value: ", ", HasSuffixSpace: false });
-      }
-    });
-
-    // Add C-variadic "..." if needed
-    if (signature.is_c_variadic) {
-      if (signature.inputs.length > 0) {
-        tokens.push({ Kind: TokenKind.Punctuation, Value: ", ", HasSuffixSpace: false });
-      }
-      tokens.push({ Kind: TokenKind.Punctuation, Value: "...", HasSuffixSpace: false });
-    }
-
-    tokens.push({ Kind: TokenKind.Punctuation, Value: ")", HasSuffixSpace: false });
-
-    // Add return type if present
-    if (signature.output) {
-      tokens.push({ Kind: TokenKind.Punctuation, Value: " -> ", HasSuffixSpace: false });
-      tokens.push(...typeToReviewTokens(signature.output));
-    }
-
-    return tokens;
+    return processFunctionPointer(type.function_pointer);
   } else if ("tuple" in type) {
     return [
       { Kind: TokenKind.Punctuation, Value: "(", HasSuffixSpace: false },
@@ -229,7 +134,6 @@ export function typeToReviewTokens(type: Type): ReviewToken[] {
       {
         Kind: TokenKind.Keyword,
         Value: type.raw_pointer.is_mutable ? "mut" : "const",
-        HasSuffixSpace: false,
       },
       ...typeToReviewTokens(type.raw_pointer.type),
     ];
@@ -242,7 +146,6 @@ export function typeToReviewTokens(type: Type): ReviewToken[] {
       {
         Kind: TokenKind.Keyword,
         Value: type.borrowed_ref.is_mutable ? "mut" : "",
-        HasSuffixSpace: false,
       },
       { Kind: TokenKind.Text, Value: elidedLifetime, HasSuffixSpace: false },
       ...typeToReviewTokens(type.borrowed_ref.type),
